refactor: 카드 브랜드 훅을 유틸 함수로 변경
